Restructure examples for package integration
