+++ .Rbuildignore
@@ -13,3 +13,4 @@
 ^odelia.Rcheck$
 ^check$
 ^gh-pages$
+^examples$
