+++ .gitignore
@@ -15,3 +15,7 @@ src/*.dll
 doc
 Meta
 *.knit.md
+
+# Rcpp auto-generated
+src/RcppExports.cpp
+R/RcppExports.R
