+++ DESCRIPTION
@@ -16,8 +16,12 @@ Depends:
     R6 (>= 2.5.1),
     dplyr (>= 1.1.0),
     tibble (>= 3.1.0)
+Imports:
+    Rcpp
+LinkingTo:
+    Rcpp
 Encoding: UTF-8
 License: GPL-2
 URL: https://github.com/traitecoevo/odelia
 BugReports: https://github.com/traitecoevo/odelia/issues
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
