+++ Makefile
@@ -1,7 +1,7 @@
 PACKAGE := $(shell grep '^Package:' DESCRIPTION | sed -E 's/^Package:[[:space:]]+//')
 RSCRIPT = Rscript --no-init-file
 
-all: compile
+all: roxygen compile
 
 # generates documentation
 roxygen:
