+++ NAMESPACE
@@ -1,2 +1,4 @@
 # Generated by roxygen2: do not edit by hand
 
+importFrom(Rcpp,evalCpp)
+useDynLib(odelia, .registration = TRUE)
