+++ R/odelia-package.R
@@ -1,0 +1,3 @@
+#' @useDynLib odelia, .registration = TRUE
+#' @importFrom Rcpp evalCpp
+NULL
